Dev2

+++ client/greenfieldt3-app/src/components/deleteItem.js
@@ -1,3 +1,4 @@
+
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 
@@ -20,7 +21,7 @@ const DeleteItem = () => {
     };
 
 
-    const handleChechboxChange = (itemId) => {
+    const handleCheckboxChange = (itemId) => {
         if (selectedItems.includes(itemId)) {
             setSelectedItems(selectedItems.filter((id) => id !== itemId));
         } else {
@@ -29,6 +30,8 @@ const DeleteItem = () => {
     };
 
     const handleDeleteSelected = () => {
+      const confirmed = window.confirm('Are you sure you want to delete the selected items?');
+        if (confirmed){
         selectedItems.forEach((itemId) => {
             axios.delete(`http://localhost:8000/item/${itemId}`)
             .then((res) => {
@@ -40,15 +43,44 @@ const DeleteItem = () => {
             });
         });
     };
-
+  }
 
     return (
-        <div>
-            <button className='deleteBtn' onClick={handleDeleteSelected}>Delete Selected</button>
-            
-        </div>
-    )
+            <div>
+      <h2 className='h2-delete'>Delete items</h2>
+      <button className='deleteBtn' onClick={handleDeleteSelected}>Delete Selected</button>
+      <table>
+        <thead>
+          <tr>
+            <th>Description</th>
+            <th>Price</th>
+            <th>Quantity</th>
+          </tr>
+        </thead>
+        <tbody>
+          {items.map((items) => (
+            <tr key={items._id}>
+              <td>
+                <label>
+                  <input
+                    type="checkbox"
+                    checked={selectedItems.includes(items._id)}
+                    onChange={() => handleCheckboxChange(items._id)}
+                  />
+                  {items.name}
+                </label>
+              </td>
+              <td>{items.description}</td>
+              <td>{items.price}</td>
+              <td>{items.quantity}</td>
+            </tr>
+          ))}
+        </tbody>
+      </table>
+    </div>
+  );
+    
 }
 
 
-export default DeleteItem;
+export default DeleteItem;
